notebooks/requirements.txt
pandas
requests
python-dotenv
tqdm
google-genai
